+++ src/focus-traversal-api-polyfill.js
@@ -1,375 +1,419 @@
+// requires DOMTraversal.
+
 (function (){
 	if (typeof window==="undefined" || window.focusManager) return;
 
-	var dom = {};
-	dom.forward = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
+	var DT = window.DOMTraversal;
 
-		var next = dom.next(element,delveShadow);
-		if (next) return dom.descend(next,delveShadow) || next;
-		else return dom.up(element,delveShadow);
-	};
-	dom.backward = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		var prev = dom.last(element,delveShadow);
-		if (prev) return prev;
-		else {
-			prev = dom.prev(element,delveShadow);
-			if (prev) return prev;
+	// array helpers
+	var flatten = function flatten(array) {
+		if (!array) throw new Error("Missing array.");
+		if (!(array instanceof Array)) array = Array.prototype.slice.call(array);
 
-			prev = element;
-			while (true) {
-				var up = dom.up(prev,delveShadow);
-				if (!up || up===document.body.parentElement) return null;
-				prev = dom.prev(up,delveShadow);
-				if (prev && prev!==document.body.parentElement && prev!==document.head) return prev;
-				prev = up;
+		return array.reduce((array,x)=>{
+			if (x instanceof Array) {
+				this.flatten(x).forEach((x)=>{
+					array.push(x);
+				});
 			}
-		}
-	};
-	dom.forwardList = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		var elements = [];
-		while (element && element!==document.body.parentElement) {
-			elements.push(element);
-			element = dom.forward(element,delveShadow);
-		}
-		return elements;
-	};
-	dom.backwardList = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		var elements = [];
-		while (element && element!==document.body.parentElement) {
-			elements.push(element);
-			element = dom.backward(element,delveShadow);
-		}
-		return elements;
-	};
-	dom.first = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		return element.firstElementChild || delveShadow && element.shadowRoot && element.shadowRoot.firstElementChild || null;
-	};
-	dom.last = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		return delveShadow && element.shadowRoot && element.shadowRoot.lastElementChild || element.lastElementChild || null;
-	};
-	dom.up = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		return element.parentElement || delveShadow && element.parentNode instanceof DocumentFragment && element.parentNode.host || null;
-	};
-	dom.down = dom.first;
-	dom.downLast = dom.last;
-	dom.next = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		if (!delveShadow) {
-			return element.nextElementSibling;
-		}
-		else {
-			var next = element.nextElementSibling;
-			if (next) return next;
 			else {
-				var parent = element.parentElement;
-				if (parent && element.parentNode instanceof DocumentFragment && element.parentNode.shadowRoot) return element.parentNode.shadowRoot.firstElementChild;
-				return null;
+				array.push(x);
 			}
-		}
+			return array;
+		},[]);
 	};
-	dom.prev = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		if (!delveShadow) {
-			return element.previousElementSibling;
-		}
-		else {
-			var parent = element.parentElement;
-			if (parent && element.parentNode instanceof DocumentFragment && element.parentNode.shadowRoot) return element.parentNode.shadowRoot.lastElementChild;
-			else return element.previousElementSibling;
-		}
-	};
-	dom.ascend = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		while(dom.up(element,delveShadow)) element = dom.up(element,delveShadow);
-		return element;
-	};
-	dom.descend = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		while (dom.down(element,delveShadow)) element = dom.down(element,delveShadow);
-		return element;
-	};
-	dom.descendLast = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		while (dom.downLast(element,delveShadow)) element = dom.downLast(element,delveShadow);
-		return element;
-	};
-	dom.descendants = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		var desc = [];
-		var e = dom.deepest(element,delveShadow);
-		while (e && e!==element) {
-			desc.push(e);
-			e = dom.forward(e,delveShadow);
-		}
-		return desc;
-	};
-	dom.ancestors = function(element,delveShadow) {
-		if (!element && !(element instanceof Node)) throw new Error("Invalid element.");
-		var ansc = [];
-		var e = element;
-		while (e) {
-			ansc.unshift(e);
-			e = dom.up(e,delveShadow);
-			if (!e || e===document.body.parentElement) break;
-		}
-		return ansc;
-	};
-
-	var tabbable = (function() {
-		// Tabbable from David Clark.
-		// https://github.com/davidtheclark/tabbable
-		//
-		// The MIT License (MIT)
-		//
-		// Copyright (c) 2015 David Clark
-		//
-		// Permission is hereby granted, free of charge, to any person obtaining a copy
-		// of this software and associated documentation files (the "Software"), to deal
-		// in the Software without restriction, including without limitation the rights
-		// to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
-		// copies of the Software, and to permit persons to whom the Software is
-		// furnished to do so, subject to the following conditions:
-		//
-		// The above copyright notice and this permission notice shall be included in all
-		// copies or substantial portions of the Software.
-		//
-		// THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
-		// IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
-		// FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
-		// AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
-		// LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
-		// OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE
-		// SOFTWARE.
-
-		var candidateSelectors = [
-			"input",
-			"select",
-			"textarea",
-			"a[href]",
-			"button",
-			"[tabindex]",
-			"audio[controls]",
-			"video[controls]",
-			'[contenteditable]:not([contenteditable="false"])'
-		];
-		var candidateSelector = candidateSelectors.join(",");
-
-		var matches =
-		typeof Element === "undefined" ? function() {} : Element.prototype.matches || Element.prototype.msMatchesSelector || Element.prototype.webkitMatchesSelector;
-
-		function tabbable(el, options) {
-			options = options || {};
-
-			var regularTabbables = [];
-			var orderedTabbables = [];
-
-			var candidates = el.querySelectorAll(candidateSelector);
-
-			if (options.includeContainer) {
-				if (matches.call(el, candidateSelector)) {
-					candidates = Array.prototype.slice.apply(candidates);
-					candidates.unshift(el);
-				}
-			}
-
-			var i, candidate, candidateTabindex;
-			for (i = 0; i < candidates.length; i++) {
-				candidate = candidates[i];
-
-				if (!isNodeMatchingSelectorTabbable(candidate)) continue;
-
-				candidateTabindex = getTabindex(candidate);
-				if (candidateTabindex === 0) {
-					regularTabbables.push(candidate);
-				} else {
-					orderedTabbables.push({
-						documentOrder: i,
-						tabIndex: candidateTabindex,
-						node: candidate
-					});
-				}
-			}
-
-			var tabbableNodes = orderedTabbables.sort(sortOrderedTabbables).map(function(a) {
-				return a.node;
-			}).concat(regularTabbables);
-
-			return tabbableNodes;
-		}
-
-		tabbable.isTabbable = isTabbable;
-		tabbable.isFocusable = isFocusable;
-
-		function isNodeMatchingSelectorTabbable(node) {
-			if (
-				!isNodeMatchingSelectorFocusable(node) ||
-				isNonTabbableRadio(node) ||
-				getTabindex(node) < 0
-			) {
-				return false;
-			}
-			return true;
-		}
-
-		function isTabbable(node) {
-			if (!node) throw new Error("No node provided");
-			if (matches.call(node, candidateSelector) === false) return false;
-			return isNodeMatchingSelectorTabbable(node);
-		}
-
-		function isNodeMatchingSelectorFocusable(node) {
-			if (node.disabled || isHiddenInput(node) || isHidden(node)) {
-				return false;
-			}
-			return true;
-		}
-
-		var focusableCandidateSelector = candidateSelectors.concat("iframe").join(",");
-		function isFocusable(node) {
-			if (!node) throw new Error("No node provided");
-			if (matches.call(node, focusableCandidateSelector) === false) return false;
-			return isNodeMatchingSelectorFocusable(node);
-		}
-
-		function getTabindex(node) {
-			var tabindexAttr = parseInt(node.getAttribute("tabindex"), 10);
-			if (!isNaN(tabindexAttr)) return tabindexAttr;
-			// Browsers do not return `tabIndex` correctly for contentEditable nodes;
-			// so if they don't have a tabindex attribute specifically set, assume it's 0.
-			if (isContentEditable(node)) return 0;
-			return node.tabIndex;
-		}
-
-		function sortOrderedTabbables(a, b) {
-			return a.tabIndex === b.tabIndex ? a.documentOrder - b.documentOrder : a.tabIndex - b.tabIndex;
-		}
-
-		function isContentEditable(node) {
-			return node.contentEditable === "true";
-		}
-
-		function isInput(node) {
-			return node.tagName === "INPUT";
-		}
-
-		function isHiddenInput(node) {
-			return isInput(node) && node.type === "hidden";
-		}
-
-		function isRadio(node) {
-			return isInput(node) && node.type === "radio";
-		}
-
-		function isNonTabbableRadio(node) {
-			return isRadio(node) && !isTabbableRadio(node);
-		}
-
-		function getCheckedRadio(nodes) {
-			for (var i = 0; i < nodes.length; i++) {
-				if (nodes[i].checked) {
-					return nodes[i];
-				}
-			}
-		}
-
-		function isTabbableRadio(node) {
-			if (!node.name) return true;
-			// This won't account for the edge case where you have radio groups with the same
-			// in separate forms on the same page.
-			var radioSet = node.ownerDocument.querySelectorAll(
-				'input[type="radio"][name="' + node.name + '"]'
-			);
-			var checked = getCheckedRadio(radioSet);
-			return !checked || checked === node;
-		}
-
-		function isHidden(node) {
-			// offsetParent being null will allow detecting cases where an element is invisible or inside an invisible element,
-			// as long as the element does not use position: fixed. For them, their visibility has to be checked directly as well.
-			return (
-				node.offsetParent === null ||
-				getComputedStyle(node).visibility === "hidden"
-			);
-		}
-
-		return tabbable;
-	})();
 
 	var history = [];
 	var historyLimit = 50;
 	var previousFocus = null;
 	var currentFocus = document.activeElement || null;
+	var traps = [];
+	var proxies = [];
+	var pending = null;
+	var id = 1000000;
 
-	var isFocusable = function(element) {
-		return tabbable.isFocusable(element);
+	/**
+	 * Returns true if a given element can receive the focus at this particular moment.
+	 *
+	 * Computation of what can and cannot receive the focus is a very complex thing.  There is
+	 * a very helpful chart at https://allyjs.io/data-tables/focusable.html that really
+	 * highlights the complexity of the issue.  THe focus-traversal-api-polyfill does its
+	 * best general case to get it right, but does not solve all use cases at this time.  It
+	 * is the hope of this module that browser manufacturers will implement the proposed
+	 * specification described at https://github.com/awesomeeng/FocusTraversalAPI in which
+	 * they will expose the internal isFocusable() method and we can all use the built-in
+	 * browser implementation and just get it right.
+	 *
+	 * In the meantime, this polyfill implementation is doing the best it can. Please let
+	 * us know if you find edge cases that dont work.
+	 *
+	 * If the given element is null or undefined or not otherwise an instance of
+	 * HTMLElement, this function will always return false.
+	 *
+	 * @param  {HTMLElement}  element
+	 * @return {Boolean}
+	 */
+	var isFocusable = function isFocusable(element) {
+		if (!element) return false;
+		if (!(element instanceof HTMLElement)) return false;
+
+		// these rules apply to the element and all its ancestors...
+		var all = DT.ancestors(element,true).reverse();
+		for (var i=0;i<all.length;i++) {
+			var e = all[i];
+			if (!e) continue;
+			if (!(e instanceof HTMLElement)) continue;
+
+			// if "inert", not focusable. Note, that inert isn't actually an HTML standard.
+			var inert = e.inert || false;
+			if (inert===true) return false;
+
+			var styles = window.getComputedStyle(e);
+
+			// if style.display = none, not focusable.
+			var disp = styles.display;
+			if (disp==="none") return false;
+
+			// if style.visibility = hidden, not focusable.
+			var viz = styles.visibility;
+			if (viz==="hidden") return false;
+
+			// if style.opacity <= 0, not focusable.
+			var op = styles.opacity;
+			op = op && parseInt(op) || 1;
+			if (op<=0) return false;
+		}
+
+		// if tabindex is -1, no focusable.
+		var ti = element.getAttribute("tabindex") || null;
+		ti = ti && +ti;
+		if (ti!==null && ti<0) return false;
+		if (ti!==null && ti>=0) return true;
+
+		// anything disabled is not focusable
+		var dis = element.hasAttribute("disabled",null);
+		if (dis) return false;
+
+		// Some elements by their very nature are focusable...
+		var tn = element.tagName;
+		if (tn==="BUTTON" || tn==="INPUT" || tn==="SELECT" || tn==="TEXTAREA") return true;
+		if (tn==="A" && (element.hasAttribute("href") || element.hasAttribute("xlink:href"))) return true;
+		if (tn==="AUDIO" && element.hasAttribute("controls")) return true;
+		if (tn==="VIDEO" && element.hasAttribute("controls")) return true;
+
+		// Any element that is contenteditable is focusable.
+		var ce = element.getAttribute("contenteditable");
+		if (ce!==null && ce===true) return true;
+
+		return false;
 	};
 
-	var focus = function(element,focusOption) {
+	/**
+	 * Sets the focus to the given element. This is functionally equivalent to doing
+	 * `element.focus()`. If no element is given, nothing will happen.
+	 *
+	 * Note that the browser does not garauntee that the focus will
+	 * be set to the given element; this is is purely a request.
+	 *
+	 * @param  {HTMLElement} element
+	 * @param  {Object} focusOption
+	 * @return {void}
+	 */
+	var focus = function focus(element,focusOption,immediately) {
 		if (!element || !(element instanceof Element)) return;
-		element.focus(focusOption||{});
+		if (typeof focusOption==="boolean") {
+			immediately = focusOption;
+			focusOption = null;
+		}
+
+		if (pending) clearTimeout(pending);
+		pending = null;
+
+		if (immediately===true) {
+			element.focus(focusOption||{});
+		}
+		else {
+			pending = setTimeout(function(){
+				element.focus(focusOption||{});
+			},0);
+		}
 	};
 
-	var hasFocus = function(element) {
+	/**
+	 * Returns true if the given element currently has the focus. This is functionally
+	 * equivalent to `document.activeElement===element;`.
+	 *
+	 * @param  {HTMLElement}  element
+	 * @return {Boolean}
+	 */
+	var hasFocus = function hasFocus(element) {
 		if (!element || !(element instanceof Element)) return false;
 		return currentFocus===element;
 	};
 
-	var orderedElements = function(container){
-		return dom.forwardList(dom.descend(container||document.body,true),true).filter(function(e){
-			return tabbable.isTabbable(e);
+	/**
+	 * Returns an array of all focusable elements within the given container.  If no
+	 * container is provided `document.body` is assumed.
+	 *
+	 * Note that in very large DOM structures this can be a significantly expensive call.
+	 *
+	 * @param  {HTMLElement|null} container
+	 * @return {Array}
+	 */
+	var list = function list(container){
+		return DT.forwardList(container||document.body,true).filter(function(e){
+			return isFocusable(e);
 		});
 	};
 
-	var first = function(container){
-		var list = orderedElements(container);
-		return list && list.length>0 && list[0] || null;
-	};
-
-	var last = function(container) {
-		var list = orderedElements(container);
-		return list && list.length>0 && list[list.length-1] || null;
-	};
-
-	var forward = function(focusOption) {
-		if (!currentFocus) return;
-
-		var target = next();
-		if (target && target!==currentFocus) target.focus(focusOption||{});
-	};
-
-	var backward = function(focusOption) {
-		if (!currentFocus) return;
-
-		var target = previous();
-		if (target && target!==currentFocus) target.focus(focusOption||{});
-	};
-
-	var next = function(element) {
-		if (!element) element = currentFocus;
-		if (!element || !(element instanceof Element)) return null;
-
-		while (element) {
-			element = dom.forward(element,true);
-			if (element && tabbable.isTabbable(element)) return element;
-		}
+	/**
+	 * Returns the first focusable element in the given container. If no container is
+	 * provided `document.body` is assumed.
+	 *
+	 * @param  {HTMLElement|null} container
+	 * @return {HTMLElement|null}
+	 */
+	var first = function first(container){
+		container = container || document.body;
+		var e = DT.bottom(container,true);
+		if (!e) return;
+		if (isFocusable(e)) return e;
+		e = next(e);
+		if (DT.contains(container,e,true)) return e;
 		return null;
 	};
 
-	var previous = function(element) {
+	/**
+	 * Returns the last focusable element in the given container. If no container is
+	 * provided `document.body` is assumed.
+	 *
+	 * @param  {HTMLElement|null} container
+	 * @return {HTMLElement|null}
+	 */
+	var last = function last(container) {
+		container = container || document.body;
+		var e = DT.bottomLast(container,true);
+		if (!e) return;
+		if (isFocusable(e)) return e;
+		e = previous(e);
+		if (DT.contains(container,e,true)) return e;
+		return null;
+	};
+
+	/**
+	 * Moves the focus forward to the next focusable element.
+	 *
+	 * @param  {Object|null} focusOption
+	 * @return {void}
+	 */
+	var forward = function forward(focusOption) {
+		if (!currentFocus) return;
+
+		var target = next();
+		if (target && target!==currentFocus) focus(target,focusOption||{});
+	};
+
+	/**
+	 * Moves the focus backward to the previous focusable element.
+	 *
+	 * @param  {Object|null} focusOption
+	 * @return {void}
+	 */
+	var backward = function backward(focusOption) {
+		if (!currentFocus) return;
+
+		var target = previous();
+		if (target && target!==currentFocus) focus(target,focusOption||{});
+	};
+
+	/**
+	 * Returns the next focusable element, but does not move the focus to
+	 * that element. If no element is provided, the currently focused
+	 * element is assumed.
+	 *
+	 * @param  {HTMLElement|null}   element
+	 * @return {HTMLElement|null}
+	 */
+	var next = function next(element) {
 		if (!element) element = currentFocus;
 		if (!element || !(element instanceof Element)) return null;
 
 		while (element) {
-			element = dom.backward(element,true);
-			if (element && tabbable.isTabbable(element)) return element;
+			element = DT.forward(element,true);
+			if (element && isFocusable(element)) break;
 		}
+
+		for (var i=0;i<proxies.length;i++) {
+			let p = proxies[i];
+			if (p.source===element) {
+				element = p.target;
+				break;
+			}
+		}
+
+		return element;
+	};
+
+	/**
+	 * Returns the previous focusable element, but does not move the focus to
+	 * that element. If no element is provided, the currently focused
+	 *
+	 * @param  {HTMLElement|null}   element
+	 * @return {HTMLElement|null}
+	 */
+	var previous = function previous(element) {
+		if (!element) element = currentFocus;
+		if (!element || !(element instanceof Element)) return null;
+
+		while (element) {
+			element = DT.backward(element,true);
+			if (element && isFocusable(element)) break;
+		}
+
+		for (var i=0;i<proxies.length;i++) {
+			let p = proxies[i];
+			if (p.source===element) {
+				element = previous(p.source);
+				break;
+			}
+		}
+
+		return element;
+	};
+
+	/**
+	 * Traps the focus within a given container. If the focus attempts to
+	 * leave the given container, focus is returned to the element that
+	 * last held the focus.
+	 *
+	 * If the given container doesn't have any focusable elements, this
+	 * call will not result in setting the focus trap.
+	 *
+	 * A `trap()` is removed by calling `untrap()` with the same container
+	 * argument.
+	 *
+	 * `trap()` cascades, so calling it multiple times causes the
+	 * trap to narrow. Removing a trap will place the browser back into
+	 * the prior trap state.
+	 *
+	 * @param  {HTMLElement} container
+	 * @return {void}
+	 */
+	var trap = function trap(container) {
+		if (!container) return;
+		if (!(container instanceof HTMLElement)) return;
+
+		var existing = traps.map(function(detail){
+			return detail.container;
+		});
+		if (existing.indexOf(container)>-1) return;
+
+		var detail = {
+			container: container,
+			last: first(container),
+			blur: null
+		};
+		traps.unshift(detail);
+
+		focus(detail.last);
+	};
+
+	/**
+	 * Stops trapping focus.
+	 *
+	 * @param  {HTMLElement} container
+	 * @return {void}
+	 */
+	var untrap = function untrap(container) {
+		if (!container) return;
+		if (!(container instanceof HTMLElement)) return;
+
+		if (traps.length<1) return;
+
+		var isfirst = traps[0].container===container;
+		traps = traps.filter(function(detail){
+			return detail.container!==container;
+		});
+		if (isfirst) focus(traps[0].last);
+	};
+
+	/**
+	 * Order one or more elements using the tabindex attribute. The order
+	 * the elements are given to this function is the order the will occur. This
+	 * will overwrite any existing tabindex on all elements given.
+	 *
+	 * If you give a number as one of the arguments, that number will be used
+	 * as the tabindex for the next element, and then incremented, and used
+	 * for the next element and so on.  If no number is given as the first
+	 * argument, an internal counter will be used instead.
+	 *
+	 * @param  {Array|HTMLElement|Number} element
+	 * @return {void}
+	 */
+	var order = function order(/*element,element,element,etc*/) {
+		var pos = null;
+		var elements = flatten(arguments);
+		elements.forEach(function(e){
+			if (typeof e==="number") {
+				pos = e;
+				return;
+			}
+			else if (e instanceof HTMLElement) {
+				let n = id++;
+				if (pos!==null) n = pos++;
+				e.setAttribute("tabindex",""+n);
+			}
+		});
+	};
+
+	/**
+	 * Sets the given source element to proxy any focus it receives to the
+	 * target element. If the source element given is not focusable, it will
+	 * be made focusable.  The target element must be focusable.
+	 *
+	 * @param  {HTMLElement} source
+	 * @param  {HTMLElement} target
+	 * @return {void}
+	 */
+	var proxy = function proxy(source,target) {
+		if (!source) return;
+		if (!(source instanceof HTMLElement)) return;
+		if (!target) return;
+		if (!(target instanceof HTMLElement)) return;
+		if (!isFocusable(target)) return;
+
+		// clear any existing proxy
+		unproxy(source,target);
+
+		var ti = source.getAttribute("tabindex") || null;
+		ti = ti && parseInt(ti) || null;
+		if (ti===null || ti<0) source.setAttribute("tabindex",0);
+
+		proxies.push({
+			source: source,
+			target: target,
+			sourceTabIndex: ti
+		});
+
+		if (currentFocus===source) focus(target);
+	};
+
+	/**
+	 * Removes an existing proxy.
+	 *
+	 * @param  {HTMLElement} source
+	 * @param  {HTMLElement} target
+	 * @return {void}
+	 */
+	var unproxy = function unproxy(source,target) {
+		proxies = proxies.filter(function(p){
+			if (p.source===source && p.target===target) {
+				if (p.sourceTabIndex!==null) source.setAttribute("tabindex",p.sourceTabIndex);
+				else source.removeAttribute("tabindex");
+				return false;
+			}
+			return true;
+		});
 	};
 
 	document.addEventListener("DOMContentLoaded",function(){
@@ -379,9 +423,22 @@
 				var path = event.composedPath();
 				if (path.length>0) target = path[0];
 			}
+
 			currentFocus = target;
-			history.unshift(currentFocus);
+
+			var trap = traps[0] || null;
+			if (trap) trap.last = target;
+
+			history.unshift(target);
 			if (historyLimit>-1) history = history.slice(0,historyLimit);
+
+			for (var i=0;i<proxies.length;i++) {
+				let p = proxies[i];
+				if (p.source===target) {
+					focus(p.target);
+					return;
+				}
+			}
 		},true);
 		document.addEventListener("blur",function(event){
 			var target = event.target;
@@ -389,7 +446,33 @@
 				var path = event.composedPath();
 				if (path.length>0) target = path[0];
 			}
+
+			var related = event.relatedTarget;
+			if (related) {
+				var trap = traps[0] || null;
+				if (trap && trap.last) {
+					var anc = DT.ancestors(related,true);
+					if (anc.indexOf(trap.container)<0) {
+						focus(trap.last);
+						return;
+					}
+				}
+			}
+
 			previousFocus = target;
+		},true);
+		document.addEventListener("keyup",function(event){
+			if (event.shiftKey && event.keyCode===9) {
+				var source = currentFocus;
+				for (var i=0;i<proxies.length;i++) {
+					let p = proxies[i];
+					if (p.source===source || p.target===source) {
+						focus(previous(p.source));
+						event.preventDefault();
+						return;
+					}
+				}
+			}
 		},true);
 	});
 
@@ -397,13 +480,18 @@
 		isFocusable: isFocusable,
 		focus: focus,
 		hasFocus: hasFocus,
+		list: list,
+		first: first,
+		last: last,
 		forward: forward,
 		backward: backward,
 		next: next,
 		previous: previous,
-		orderedElements: orderedElements,
-		first: first,
-		last: last
+		trap: trap,
+		untrap: untrap,
+		order: order,
+		proxy: proxy,
+		unproxy: unproxy
 	};
 	Object.defineProperty(window.focusManager,"currentlyFocused",{
 		get: function() {